requirements.txt
python-dotenv == 1.0.1
Flask == 3.0.3
requests == 2.31.0
Pillow == 10.3.0
gunicorn == 22.0.0
flasgger == 0.9.7.2.dev2
